LayoutTests/canvas/philip/tests/2d.line.width.valid-expected.txt

< [index] >
2d.line.width.valid
Setting lineWidth to valid values works
References: 2d.lineWidth.set 2d.lineWidth.get
Actual output:
Passed
